Fix build on arm32 (#1108)
On arm32, size_t and long are 32 bits in size, exposing these bugs.

+++ src/security/PeerOptions.cc
@@ -334,7 +334,7 @@ static struct ssl_option {
 #endif
 #if SSL_OP_ALL
     {
-        "ALL", (long)SSL_OP_ALL
+        "ALL", SSL_OP_ALL
     },
 #endif
 #if SSL_OP_SINGLE_DH_USE
